Support scroll bars in search result table.
Finally the result table shows automatically horizontal and vertical
scrollbars, when additional columns are selected. For achieving this
behavior the parent composite of the result table gets an
FillLayout. After the result table is generated the layout method is
called:

 `tableParentComposite().layout()`

This calls the layout manager. This is normally done, if the parent
composite is resized, but has to be fired manual, after generating the
table. It repaintes all underlying composites, if a layout manager is
set.

(VN-1083)

git-svn-id: svn+ssh://svn.verinice.org/data/subversion/svnroot/TRUNK@4760 4c0145ca-515c-4d1a-a6bc-f9bd594c3410