Implement highlighting in search table view.
The highlighted fields API is not used right now, because the complete
text fragment is wrapped with the html tag. I have no idea, if this is
a bug or depends on the way how the index is created. For now the
query text of the user is used for matching the strings in the table
cells and colorize them.

This is inconvenient, because if the Elasticsearch query is changed
(for example from case insensitive to case sensitive) verinice would
continue to mark Strings case insensitive.

(VN-1083)

git-svn-id: svn+ssh://svn.verinice.org/data/subversion/svnroot/TRUNK@4786 4c0145ca-515c-4d1a-a6bc-f9bd594c3410