Add blocking index method for elastic search.
For testing purposes this method is really needed, because junit test
can only run, if an index is really created.

As a side effect there is also a correct measurement for the duration of
the indexing process implemented, also for the non blocking index method.

Note that the blocking indexing method is as fast as the non blocking
method.

(VN-1082)

git-svn-id: svn+ssh://svn.verinice.org/data/subversion/svnroot/TRUNK@4885 4c0145ca-515c-4d1a-a6bc-f9bd594c3410