Add Greenbone Perspective and Cheat Sheets
vn-1150

Add a new link with a Greenbone logo to the welcome screen. The link opens a new Greenbone perspective which is identical to the ISM perspective but when opened via the welcome screen link it launches with a new cheat sheet opened on the right side of the application window which is supposed to explain how to connect Greenbone Security Manager to Verinice. In this commit however the cheat sheets are only scaffolds without actual content.

git-svn-id: svn+ssh://svn.verinice.org/data/subversion/svnroot/TRUNK@5049 4c0145ca-515c-4d1a-a6bc-f9bd594c3410