save FinishedRiskAnalysisList after initialize
VN-975

on the first page of the itgs risk analysis  wizard, all gefaehrdungen
of the bausteine which are children of the selected cnatreeelement are
getting selected. those needs to be saved after initialization,
otherwise the selection will not be passed to the second page (only on
the first try (forward/back/forward within wizard triggers several
saves)). this commit fixes that bug, which also affects VN-1309 (if no
gefaehrdung is in the list, the next button is grayed out)