Remove useless IPropertyChangeListener from BsiModelView
vn-1324

The IPropertyChangeListener closed the IT Baseline Protection model when
the settings for the database changed. But this only produced a
nondescript error message in the GUI if the user wanted to make changes
to the model without restarting the client. The changes still got
written to the *old* database. This doesn't help the user at all.

Instead, the user should be forced to restart a client after changing
properties of the database setup, but that's a different issue.