Exports all references defined by the SNCA.xml.
VN-456

References have been stored as database ids of entity objects in the
entity table directly. The VNA export has not taken into account that
the entities have a different database id in foreign database the VNA
file is imported into, so this approach is insufficient.

The new approach is to map the entity database id to the external id
during the export process. The import makes a reverse mapping from
external id to the actually entity database id.

This solution is backwards compatible, since verinice < 1.12 exported
and imported useless values as references and the UI behaves fault
tolerant.