Migrate static html pages to jsf with language switch
VN-1323

Every page is not either in english (error pages) or can be changed
between german and english

The layout of all pages is adapted and unused files are removed

note: "nav.xhtml" and the error pages were deleted by mistake and then
restored from local history. So for some reason it is marked as deleted
and added.