Append minor minor version number in verinice about dialog.
There is also a version number in the about dialog. Often customer are
asked, which client version they exactly use and when they are it is
easy for them to look it up in the client about dialog.

So with the minor number at the end, it is possible for the end user to
see the different between hotfixed clients.