Squashed 3 to one commit, due to merge conflicts
VN-1342

exception occurs if a it-network below imported-objects containing a
risk analysis is enqueued for deletion
prevent throwing exception on deletion of GefaehrdungsUmsetzung

VN-1342

moved deletion (in case of IT-Network) of risk analyses after removal of
configuration instances, afterwards added deletion of manually added
instances of GefaehrdungsUmsetzung
fix deleting riskanalyses containing elements

VN-1342

last commit only fixes deleting it networks underneath the imported
object folder, which means the error described in the issue, still
appears when the user tries to delete the whole imported objects folder,
or a category containing an element containing a risk analyses. this is
fixed by this commit.