Make FinishedRiskAnalysisLists copyable
VN-1250

-Resolve wrong canContain of FinishedRiskAnalysis
Before the FinishedRiskAnalysis could contain BausteinUmsetzung,
Baustein, LinkKategorie and FinishedRiskAnalysis as well as
GefaehrdungsUmsetzung. This is now prevented and only
GefaehrdungsUmsetzung is allowed
-
Remove Blacklist and other restriction from GroupTester
Remove the copy-constraint for RisikoMassnahmenUmsetzung,
FinishedRiskAnalysis and GefaehrdungsUmsetzung. Add a constructor to
RisikomassnahmenUmsetzung to make copying possible.

-Add copying of FinishedRiskAnalysisLists to CopyCommand
The Lists needed for the content of the RA_Wizard are now copied as
well. In general the FinishedRiskAnalysis-element is now handled
differently from the other elements.