add custom classloader for reporting
VN-1394

implemented custom classloader that is set to bsh interpreter to
whitelist code that can be executed via beanshell. cleaned up 3
rptdesign files as a sideeffect (of building the whitelist).
also renamed java.policy to verinice.policy, but it seems like that this
is not needed anymore to fulfill requirements of this issue.