Introduce a simple schema validator for verinice.
VN-1412

Every verinice instance has knowledge about its supported schema version
from now on. If verinice imports a vna/xml file, it is checked if this
file is compatible with current verinice version.

Whenever the data are exported verinice wrote the current schema version
into the xml output and additionally the schema versions which are
compatible to the current schema.