add a sandbox for report generation
VN-1394

execution of beanshell and javascript (rhino) needs to be sandboxed
within report-templates, so that no unauthorized code could be executed
while using verinice. to solve this, the report engine runs/renders the
report within a thread that runs in the context of a special
securitymanager, and a custom classloader is passed to the beanshell
interpreter additionally.