Implement code review remarks
VN-1441

Add some class documentation and remove forgotten comments.
In addition a line which was added in the acceptance test and wasn't
committed is now added to ensure drop support to the BSIModelView.
Finally the remote calls for the uuid objects are now bundled in a new
command which loads all elements by a uuid list.