pass reason for securityexception to user(/ui) lvl and refactor report-ui-classes
VN-1394

unwraps the reportsecurityexception (if appeared) out of the birt
"engineException" and hands it up to ui-lvl to inform the user with an
error dialog. furthermore the code for the differen actions enabling
calling a report / the report dialog gets unified by an abstract class
and some refactoring. Additionally the texts for the
ReportSecurityException are now i8lnized. Query and
ReportSecurityManager making use of ReportClassLoader again, which seems
to be necessary.