Make restriction of service methods in extra files.
VN-1393

Background: Spring security is switched on, if there it is referenced in
the root <beans> tag. This is unwanted when running the verinice server
embedded, so we have to exclude some beans from the verinice-common.xml
file in extra files, when we want to use the security namespace for the
method  interceptor.