Use InheritableThreadLocal for VeriniceContext.
The main UI-Thread is initialized with a ThreadLocal-Map for providing
the spring context which contains verinice services to the rcp
framework. The ThreadLocal is not copied to child threads, so the
verinice developer were used to load the VeriniceContext manually by
this line of code, when they are working with services within threads:

    Activator.inheritVeriniceContextState();

The switch to the InheritableThreadLocal makes this call unnecessary and
the concurrent ui programming less error prune.

One important remark: On server side there is a similar method:

    ServerInitializer.inheritVeriniceContextState

This method is used on server side code in services and we need further
investigations if this also helps there.