consider attachments on element copy
VN-578

add a (boolean-)preference that enables to consider cnatreeelement
referencing file attachments. On database-layer, a second entry
describing the attachment (a real hardcopy) will be created. Furthermore
the code to create an instance of an AttachmentFile gets encapsulated in
a Factory, to remove code duplication in several places throughout the
app.