add licenseManagement relevant properties to sernet.hui.common.connect.Property
VN-1536

for the implementation of the license-management-feature the class
sernet.hui.common.connect.Property needs to be extended by to attributes
that stores the information if a property is restricted by a license,
and what license is restricting the property. This attributes needed to
get mapped by hibernate (for derby, postgres and oracle) and needed to
be considered by import and export (vna), and keep compatability to the
old (licensefree) vna format also (downwards-compatability). all done by
this commit