First draft of ltr rendering re-implementation
VN-1528

Implements a new approach for creating a link table in verinice. There
is missing the filtering on edges on the output of edge
properties (called links in verinice). That is why this implementation
is not switched on by default.

There is one line in the ExportLinkTableHandler:76 which is commented
out. If you want to test this implementation, you'll have to remove this
comment.

- Refactoring:
  - Makes no sense to use generics for the adapter pattern.
  - Refactor train wreck.
  - Move the creation of header columns to service.