Refactor crypto from osgi to spring
VN-1538

Plugin sernet.verinice.encryption (used via OSGI by client only) gets
removed and encryption service gets moved to service, with own
interface, and is accessible via applicationcontext via spring.
Licensemanagementservice does make use of this, and its unittest as
well. Please note, not all unit-tests and methods of the service already
using the crypto-methods, so its obvious some of the @test-marked
methods failing currently. this is a really deep change in our general
architecture and needs to be tested with care and in detail. 
Furthermore this commit is not fully functional xet, it is a kind of a
backup commit. Do not use this commit in production environmments.