Make the port of internal server configurable.
VN-1268

If a user of verinice has already assigned the standard port 8080 to
another process, it is now possible to change the port with an entry in
the verinice.ini file:

    -Dorg.osgi.service.http.port=8181

This changes the port for the internal server. The client also reads the
system property and modifies its internal server URI according to this
port.