Add LDAP search to ldap authentication.
VN-1292

If in `verinice-ldap.properties` the value `ldap.groupfilter` is set,
a ldap search is executed with the username on order to retrieve the
right `dn`. This approach makes it possible that user can be
authenticated by there `uid`, although they are store in different
LDAP sub trees:

	dn: uid=user_1,dc=foo,dc=example,dc=com
	dn: uid=user_2,dc=bar,dc=example,dc=com

This configuration prevent ldap authentication by the simple prefix
(pattern) approach by the `ldap.prefix` variable. So in this case,
leave the `ldap.prefix` variable empty and set the `ldap.groupfilter`:

	ldap.groupfilter=dc=example,dc=com

Afterwards verinice will query for a `uid` in this tree and retrieve
the right `dn`.