Fix method call of secured task service in web frontend.
(VN-1393)

The task view in the web frontend list task for the currently logged
in user. The old call works with a parameter object which contains the
user name of the user for which the taks should be displayed. The
working call for this without additional action rights is now:

	sernet.verinice.interfaces.bpm.ITaskService.getCurrentUserTaskList()