add behavior for multioptionselectwidget
VN-1573

if a property in a multioptionselectwidget is dependet, the behaviour
should react also if some more options of the widget are select, and not
only if the option, definied in the snca depends attribute, is selected
all alone. The new class defines that behaviour by calling
String.contains() on the control text (representing the checked values)
checking for the value defined in the snca depends attribute.