prevent reports from showing empty tables
VN-1511

both reports, itgs-a7 and realisierungsplan, should only show table
entries for bausteine that has children of type massnahme that has an
umsetzungsstatus-value that not equals "ja" or "entbehrlich". up to now
the reports were showing all bausteine / massnahmen. new versions
prevent to show massnahmen with umsetzungsstatus "ja" or " entbehrlich"
and prevent to show bausteine that only contain massnahmen with one of
that umsetzungsstatus-values.