add a update news dialog
VN-1588

add the possibility to show (update) news from an external resource,
stored in json and displayed in html and in english and german as well.
users have the possibility to decide if he wants to install the offered
update and to switch of displaying the notification. this is a
standalone-only feature, the implementation of the interface on the
serverside is just a dummy implementation