Support D&D for BausteinUmsetzung -> ISO-Person
VN-1639

Creating a HuiRelation via Drag and Drop between the ISM Perspective
and the BSI-Grundschutz Persective had not worked for
BausteinUmsetzung (CnaTreeElement instances of a Baustein) since it
was not a valid source in DragSourceListener and DNDHelper.

This patch uses the IBSIStrukturElementTransfer class in order to
enable Drag and Drop Support. This seems to be necessary, because
there is bug in the RCP GTK component, because the BSIModelView have
already enabled drag support for `Bausteine`:

    Transfer[] dragTypes = new Transfer[] {
    IBSIStrukturElementTransfer.getInstance(),
    BausteinUmsetzungTransfer.getInstance() };

    int operations = DND.DROP_COPY | DND.DROP_MOVE;
    viewer.addDropSupport(operations, dropTypes, new BSIModelViewDropListener(viewer));
    viewer.addDragSupport(operations, dragTypes, new BSIModelViewDragListener(viewer));

The weird thing is that you can change the order of drag types and the
creating of relations between ISM objects and `Bausteine` will
work. After digging deep in the RCP 3.7 source code, I came at this
place:

    void dragGetData(long /*int*/ widget, long /*int*/ context, long /*int*/ selection_data,  int info, int time){
        if (selection_data == 0) return;
        GtkSelectionData gtkSelectionData = new GtkSelectionData();
        OS.memmove(gtkSelectionData, selection_data, GtkSelectionData.sizeof);
        if (gtkSelectionData.target == 0) return;

        TransferData transferData = new TransferData();
        transferData.type = gtkSelectionData.target;
        transferData.pValue = gtkSelectionData.data;
        transferData.length = gtkSelectionData.length;
        transferData.format = gtkSelectionData.format;

        DNDEvent event = new DNDEvent();
        event.widget = this;
        event.time = time;
        event.dataType = transferData;
        notifyListeners(DND.DragSetData, event);

        if (!event.doit) return;
        Transfer transfer = null;
        for (int i = 0; i < transferAgents.length; i++) {
            Transfer transferAgent = transferAgents[i];
            if (transferAgent != null && transferAgent.isSupportedType(transferData)) {
                transfer = transferAgent;
                break;
            }
        }

        ...
    }

This code always choose the wrong transfer agent because the object
`gtkSelectionData` is always initialized with the from `target`
value. It is always the `Id` of the transfer agent which is the first
one in the first code fragment. I have no idea why.

So this approach is to make a special treatment of `Baustein` in the
`IBSIStrukturElementTransfer` and simply ignore the strange behavior
of RCP/GTK.