add possibilty to set property displayed in browser view via snca
VN-1569

Every CnATreeElement should get a property that will be displayed in the
browser view (html). To avoid the need of changing CnATreeElement or
writing a new interface for this, this commit introduces a new
xml-attribute to HuiProperty. The datatype is boolean and false by
default, if it is set, the first property (in chronological order) that
has this property set will be the one that is going to be displayed in
the browser view if an instance of the specified HuiEntity is selected
in the TreeView or opened in an editor. In case of no property gets
assigned with the new attribute, the old logic will take place, so there
is no need to change the data and the old behaviour will still be in
place.