add usage of de-/encryption to all methods of LicenseManagementService
VN-1538

all usage of getters from LicenseManagementEntry in both of the
implementations if LicenseManagamentService needs to access the de- and
encrypt-methods of the encryptionservice, to ensure the correct data is
read from and written to the database. There is still a decision to
take, how to handle the storage of granted licenseId in the
Configuration-Objects, de- or encrypted. When decision is made, some
changes are needed in the service-implementations.