Revise Report Query Editor GUI
vn-1641

This changes the Report Query Editor GUI in the following way: instead of
having move-up, move-down and remove buttons in every row of the editor,
there is a tool bar now with those buttons. Every row has a radio button
and the buttons in the toolbar operate on the selected row if
applicable.