Fix the parent relations operator.
VN-1187

Replace old commons-collection in test launcher.
------------------------------------------------

Introduce by fixing the serialization venerability with commit
updated. Done this with following snippet:

	find -name *.launch | xargs sed -i s/commons-collections-3.2.jar/org.apache.commons.collections_3.2.2.jar/

Edit LTR output tests.
----------------------

Lines should be filled up when an element in the tree has at least one
parent in the data set.

Print parent relations in the same row.
---------------------------------------

A query is parsed as a simple tree.

	"asset<assetgroup<assetgroup.assetgroup_name AS Title",
	"asset<assetgroup.assetgroup_name",
	"asset.asset_name AS Title",
	"asset/control.control_name AS Title"

would be parsed to:

                    asset
                   /     \
              (<) /       \ (/)
                 /         \
             assetgroup  control
                /
           (<) /
              /
         asseetgroup

Since the control is another path from the asset, it was printed
within an extra row. This is not right the behaviour since on the left
path the parent operator is used. This entries will be printed
together with the control in a single row.

                        asset
                       /  |  \
                      /   |   \
                     /    |    \
                    /     |     \
                   /      |      \
              (<) /       | (/)   \ (/)
                 /        |        \
             assetgroup control  scenerio
                /
           (<) /
              /
         asseetgroup

For this example query all found assetgroups are appended to controls
and scenarios.