Use credentials for LDAP-search.
VN-1666

The LDAP search mechanism was implemented without any credentials for
the search layout. That means that the LDAP search was only
success full, if the search user does was not restricted by a password.

New properties are introduced in the verinice-ldap.properties:

	# User for searching principals
	ldap.search.user=

	# Password for user search.
	ldap.search.password=

These values have to be defined if a user filter and group filter is defined,
which are is the necessary condition that verinice uses the LDAP search.