fix security issue with task-view
VN-1598

opening the taskview, an unpriviliged user produces an error because of
inadequate security-settings. The Combobox comboAccount of the taskview
shows the current user, or all users (if actionId "TASKSHOWALL" is set)
and with that the accountservice is called. The method used for this so
far was "AccountService.findAccounts" which only was whitelisted for
being used by users with permission of actionId "ACCOUNTSETTINGS", which
is not part of "user-default-profile", and should not be. So this commit
adds the actionId "TASKSHOWALL" to the whitelist of method
AccountService.findAccounts in case of the user needs to see / load all
available accounts. If the user only needs to see his own account, he
needs (by refactoring) to access AccountService.getAccountByName. As he
has the actionID TASKVIEW anyways (to access the taskview), it is added
to the whitlelist of AccountService.getAccountByName.