Fix checkbox values in web editor.
Problem seems to be introduced through the task release workflow, which
maintains a seperate map for changed values. This map can only handle
strings. So this fix looks for boolean values and handles them properly
but at the end of the day this is a quick fix and it should be done
properly one day.