Initialize the syncRequest in pure xml when called from soap.
VN-1719

The constructor of the SyncCommand called from the soap server need to
set the syncRequest object otherwise the isCompatible(VnaSchemaVersion
vnaSchemaVersion) method tries to unmarshal the SyncRequest which fails
with a npe as no veriniceXml is set. This makes sense as the SyncRequest
object is already instantiated by the soap service.