Remove the v-designer spring context as we use the springClient one
VN-1716

To use the SNCA.xml from the server we now use the springClient bean
context and get the ObjectModelService as service from the OSIG
environment. The bundelUrlLoader the serviceFactory and and the
spring-vdesinger.xml are obsolete therefore removed. The connection
properties are taken from the datasource just as the old
dataset(skripting) does. 
Reorganized imports for the touched classes.