remove the todos update the test
VN-1708

Introduced a new report property 'multiple_root_objects' which is false
per default. When set to true the generate Report option 'scope' is
extended by an entry 'all' which generates the report for all
elements/scopes.
This property can be set in the AddReportToDepositDialog. 
Update the test to reflect the new property and remove and addressed and
removed the todos.