Validate if snca is loaded.
VN-1716

Check via the object model service if a connection to the server can be
established. This is more a hack than a solution. But the whole
datasource concept is broken by design, since we initialize the spring
context for the whole application and not per datasource. To fix this
will take much more effort and seems to be out of context in this issue.

Also the initializing process of HuiTypeFactory is extended with a new
HuiTypeFactory which indicates the failure of the initializing.