handle RiskOwnSafeguards in ObjectBrowser like before implementation of licence-Management
VN-1753

the text that is displayed (by default) in the objectbrowser for objects
of type RisikoMassnahmeUmsetzung is stored in RisikoMassnahme. That
means, it can not be considered / handled the generic way, implemented
for license-management. So this commit fixes that and resets the
treatment of RisikoMassnahmeUmsetzung for displaying them within the
Object-Browser