Set verince context in every request.
VN-1771

If no context is available or is empty the thread local with the
verinice services is reinitialized. The call is shielded because the
code is synchronized and that is why it is quit expensive to
reinitialize the context.

On the long run we should not use the method:

	sernet.gs.service.ServerInitializer.inheritVeriniceContextState()

It is not clear why the context is not available and why we use a
`ThreadLocal` for accessing Spring Beans from outside, because the
`ServerInitializer` already contains a reference to the String Bean
map and every bean follows the singleton pattern, which means it is
not nessary to use a `ThreadLocal` anyway.