Introduced a mechanism to map command to the RightsServer.
vn-1516

In the HibernateCommandService a Map<CommandClass,Set<String>> is
introduced configured by the command-actionid-mapping.xml spring bean
file. This map is used to defines extra constrains on the execution of a
command. All defined actionid must be enabled by the rightsServerHandler
to execute the command. If no no set is defined the command is allowed.