Store option value instead of option name in task.
VN-1833

If a value of combobox with numerical values is changed in the
webapplication and the task release workflow is active, there has to be
stored the numerical representation and not the message
string. Otherwise we end with a class cast exception when restoring this
value in the task view.