Gaurd and reset changed properties map for tasks.
VN-1833

Since EditBean is session scoped, the tracking map for changed element
contains all changes, which are done to several elements. This is a
problem when completing a tasks and the server tries to write back
properties for elements which are not defined for these elements.

So this solution fills only the changed properties map when we are in
the "Change Task Request" context and reset the map every time a new
Tasks is opened.