Turn off automatic detection of file upload mode.
VN-1819

Since we are using servlet API 2.5 for our web application we force
primeface to use the "commons" file uploader. Otherwise and under
circumstances primefaces uses native (e.g with tomcat 6 under centOs),
which does not work properly.

With this method we are save to run our web application in every
container server without updating any server container libraries.

Because of all this we delete the additional libraries from server and
the legacy documentation file which are not needed for the tomcat
instance.