Parse date correctly for task compare view.
VN-1838

This bug was discovered during checking the property dependency
behaviour in the webapplication and commit is not directly related to
the depends mechanism, but it seems, that the changed dates within the
task release workflow context can not applied to a CnATreeElement and
the element is corrupted afterwards. So I have decided, to fix this as
soon as possible.