Refresh CatalogView after successful delete
VN-1796

Trigger the refresh is quite tricky since the DeleteHandler looks for
the proper root model of the deleted element. The CatalogModel is an
alternative root model, hence the DeleteHandler can't find out which
root model to refresh. There the DeleteHandler expects the default root
model as before but in addition also always refreshes the CaralogModel.