Replace usage of Qualifier String with ENUMs
VN-2142

The ENUM values are stored as global unique hui property ids. This id
shall only be used by corresponding hui entity.  This also means, that
this key should not be used by the API to localize the representation of
the ENUMs. Instead the ENUMs handle their localization on their own.