Do a refactoring to improve the migration command
VN-2037

Create a constant for the system property to disabled migration.
Seperate the code which skips the migration.
Introduce a method which initializes elemements without using them.
Log the number of processed requirements and threats.
Output DB ID of elements in tree for debugging.
Output DB ID in toString method of CnATreeElement.
Split the migration code into 3 commands
Transfer the code from migration command to migrate requirements and threats to a 2 new commands.
Improve unit test