Configure the internal Browser
VN-2460

Disabling the context menu & a workaround to disabling internet access &
dnd. SetText sets the URL to file:/// or about:blank, that causes Reload
(or Back + Forward) to break. As a workaround, the context menu was
disabled. For various resasons the Browser now only is allowed to access
about:blank (SetText) and the logo.